Add files via upload
Computer made a snake and random coordinate leaf using TURTLE. The snake start mave and we press the left key snake move to the left this type right, up and down. Snake reached the leaf then the leaf hidden and increase the snake size. Snake reach the out of the screen then game over.